backend/py_requirements.txt
asgiref==3.8.1
Django==4.2
django-cors-headers==4.7.0
djangorestframework==3.15.2
Faker==36.1.0
sqlparse==0.5.3
tzdata==2025.1
